requirements.txt
langchain==0.0.163
openai==0.27.6
pinecone_client==2.2.1
streamlit==1.22.0
